+++ CHANGES.rst
@@ -14,6 +14,9 @@ Changelog
   [gbastien]
 - Fixed adding a MeetingConfig TTW, set correct default values.
   [gbastien]
+- Fixed activate correct `portal_tab` while using grouped configs and several
+  MC start with same id.
+  [gbastien]
 
 4.2b4 (2020-10-14)
 ------------------