+++ CHANGES.rst
@@ -19,6 +19,9 @@ Changelog
   from first item. Manage also when attendee is `excused/absent` then
   `non attendee` and so still not present.
   [gbastien]
+- Fixed activate correct `portal_tab` while using grouped configs and several
+  MC start with same id.
+  [gbastien]
 
 4.2b4 (2020-10-14)
 ------------------