+++ CHANGES.rst
@@ -22,6 +22,9 @@ Changelog
 - Fixed activate correct `portal_tab` while using grouped configs and several
   MC start with same id.
   [gbastien]
+- Use position `bottom` to display tooltipster `usersGroupInfos`
+  to avoid screen overflow.
+  [gbastien]
 
 4.2b4 (2020-10-14)
 ------------------