+++ CHANGES.rst
@@ -11,7 +11,9 @@ Changelog
   - losing information (`fullname`) on elements the user interacted with;
   - loading the application and maybe break it as `local_roles` are recomputed
     on every existing elements by Plone when deleting a user.
+
   [gbastien]
+
 - Fixed adding a MeetingConfig TTW, set correct default values.
   [gbastien]
 - Display group `Administrators` members on the MeetingConfig view.